% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_by_correlations.R
\name{annotate_by_correlations}
\alias{annotate_by_correlations}
\title{RETROFIT matching algorithm}
\usage{
annotate_by_correlations(sc_ref, K, decomp_w, decomp_h)
}
\arguments{
\item{K}{integer: The number of cell types to be selected}

\item{decomp_w}{Matrix(GeneExpressions, Components): Decomposed w matrix}

\item{decomp_h}{Matrix(Components, Spots): Decomposed h matrix}

\item{ref_w}{A Matrix or Array with two dimensions (GeneExpressions, Cell types).}
}
\value{
A list of
\itemize{
\item w
\item h
}
}
\description{
Match cell types based on correlations with reference. decomp_w   between matching algorithm description
}
\examples{
K = 8
ref_w=read.csv(paste("../data", "sample_sc_ref.csv", sep="/"), row.names = 1, check.names = FALSE)
decomp_w=read.csv(paste("../data/sample_results", "sample_x__decomp_w.csv", sep="/"), row.names = 1, check.names = FALSE)
decomp_h=read.csv(paste("../data/sample_results", "sample_x__decomp_h.csv", sep="/"), check.names = FALSE)
result = RetrofitMapByCorrelation(sc_ref=ref_w, 
                                 K=K,
                                 decomp_w = decomp_w,
                                 decomp_h = decomp_h)
}
\seealso{
papers reference
}
