% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{RETROFIT decomposition algorithm}
\format{
TestDecomposeData
A list supporting decompose testing
\describe{
\item{x}
}
}
\usage{
decompose(
  x,
  L = 16,
  iterations = 4000,
  lambda = 0.01,
  seed = NULL,
  alpha_w_0 = 0.05,
  beta_w_0 = 1e-04,
  alpha_h_0 = 0.2,
  beta_h_0 = 0.2,
  alpha_th_0 = 1.25,
  beta_th_0 = 10,
  kappa = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{x}{matrix or array with dimension (GeneExpressions, Spots). This is the main spatial transciptomics data.}

\item{L}{integer (default:16)    The number of components to be decomposed}

\item{iterations}{integer (default:4000)  The number of maximum iterations to be executed}

\item{lambda}{double  (default:0.01)  Background expression profile control}

\item{seed}{double  (default:NULL)  Random variable seed in case the output should be deterministic}

\item{alpha_w_0}{double  (default:0.05)  Variational initial parameter for vector alpha_w}

\item{beta_w_0}{double  (default:0.0001)Variational initial parameter for vector beta_w}

\item{alpha_h_0}{double  (default:0.2)   Variational initial parameter for vector alpha_h}

\item{beta_h_0}{double  (default:0.2)   Variational initial parameter for vector beta_h}

\item{alpha_th_0}{double  (default:1.25)  Variational initial parameter for vector alpha_th}

\item{beta_th_0}{double  (default:10)    Variational initial parameter for vector beta_th}

\item{kappa}{double  (default:0.5)   Learning rate factor}

\item{verbose}{boolean (default:FALSE)}
}
\value{
A list of decomposed vectors that contains
\itemize{
\item w:             2d array with GeneExpressions, Components
\item h:             2d array with Components, Spots
\item th:            an array with Components
\item durations:     (verbose) durations vector (unit: second)
\item relative_error:(verbose) errors with pre-defined norm vector
}
}
\description{
Receiving the input with 2d spatial transcriptomics matrix,
the function returns factorized {W, H, Theta}.
This function fulfills Structured Stochastic Variational Inference Algorithm for RETROFIT.
Since exact Bayesian inference is infeasible and considering the large number of spots and genes,
variational inference was adopted to approximately estimate the parameters in performant manner.
}
\examples{
data("TestDecomposeData")
x   = data$extra5_x
res = retrofit::decompose(x, L=16, iterations=10, verbose=TRUE)
W   = res$w
H   = res$h
TH  = res$th
}
\seealso{
papers reference
}
