% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_metric.R
\name{delta_metric}
\alias{delta_metric}
\title{RETROFIT delta metric}
\usage{
delta_metric(cor_matrix)
}
\arguments{
\item{cor_matrix}{matrix or array with dimension (CellTypes, Parameters).
For example: cor_matrix =
0   0.01   0.05
Endothelium  0.541  0.581  0.167
Epithelial   0.756  0.725  0.684
Muscle       0.881  0.869  0.780

Here, there are 3 cell types and 3 different parameter settings.
The values of the matrix are the correlation between the estimated gene expression profile via RETROFIT,
and the corresponding cell-type specific marker gene expressions at different parameter settings.}
}
\value{
A dataframe that contains
\itemize{
\item parameters: different parameter settings provided in the input
\item delta:      delta metric for each parameter setting
}
}
\description{
Receiving the input with correlation matrix for different cell types (rows) at different parameter settings (columns),
the function returns the delta metric.
}
\seealso{
Algorithm 1 in paper
}
