% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lambdaCorrelationData}
\alias{lambdaCorrelationData}
\title{Lambda correlation data}
\format{
\subsection{\code{lambdaCorrelationData}}{

A data frame with the correlation matrix between lambda values and cell types.
\describe{
\item{cor_matrix}{8 x 6 Cell types x Lambda values}
}
}
}
\source{
Roopali Singh, Xi He, Adam Keebum Park, Ross Cameron Hardison, Xiang Zhu, Qunhua Li. RETROFIT: Reference-free deconvolution of cell-type mixtures in spatial transcriptomics, Preprint Forthcoming (2023)
}
\usage{
data(lambdaCorrelationData)
}
\description{
A dataset with the correlation matrix of the lambda parameter and cell types supporting parameter selection
}
\details{
\itemize{
\item lambdaCorrelationData
}
}
\keyword{datasets}
