% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.retrofit}
\alias{predict.retrofit}
\title{Predict using retrofit}
\usage{
\method{predict}{retrofit}(retrofit, new_x)
}
\arguments{
\item{retrofit}{An object of class `retrofit` obtained using the `retrofit()` function}

\item{new_x}{A numeric vector containing new scores to retrofit using the above object}
}
\description{
Given a set of predictions and truth for two
models, calibrate the result of the second to yield the
same expected value as the result of the first.
}
