% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readmeData}
\alias{readmeData}
\title{Readme}
\format{
Includes input x
}
\usage{
data(readmeData)
}
\description{
A dataset supporting readme example
}
\details{
\itemize{
\item readmeData
}
}
\keyword{datasets}
