% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readmeObj}
\alias{readmeObj}
\title{Readme}
\format{
Includes input x
}
\usage{
data(readmeObj)
}
\description{
A dataset supporting readme example
}
\details{
\itemize{
\item readmeObj
}
}
\keyword{datasets}
