% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit.R
\name{retrofit}
\alias{retrofit}
\title{Retrofit}
\usage{
retrofit(old_pred, old_truth, new_pred, new_truth, method = "gpava")
}
\arguments{
\item{old_pred}{Vector of predictions of the old model}

\item{old_truth}{Vector of true results of the old model}

\item{new_pred}{Vector of predictions of the new model (the one to be calibrated)}

\item{new_truth}{Vector of true results of the new model}

\item{method}{Method used to perform the calibration}
}
\value{
An object of class `retrofit` that can be used to
convert the new model score to match the old model score.
}
\description{
Given a set of predictions and truth for two
models, calibrate the result of the second to yield the
same expected value as the result of the first.
}
