% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_by_markers.R
\name{RetrofitMapByMarkers}
\alias{RetrofitMapByMarkers}
\title{RETROFIT matching algorithm}
\usage{
RetrofitMapByMarkers(ref_marker, K, decomp_w, decomp_h)
}
\arguments{
\item{ref_marker}{Key-value list: A dictionary of key: cell type, value: GeneExpression list}

\item{K}{integer: The number of cell types to be selected}

\item{decomp_w}{Matrix(GeneExpressions, Components): Decomposed w matrix}

\item{decomp_h}{Matrix(Components, Spots): Decomposed h matrix}
}
\value{
A list of
\itemize{
\item w
\item h
}
}
\description{
Match cell types based on correlations with reference. decomp_w   between matching algorithm description
}
\examples{
K = 8
decomp_w=read.csv(paste("../data/sample_results", "sample_x__decomp_w.csv", sep="/"), row.names = 1, check.names = FALSE)
decomp_h=read.csv(paste("../data/sample_results", "sample_x__decomp_h.csv", sep="/"), check.names = FALSE)
ref_marker_d=read.csv(paste("../data", "sample_ref_marker.csv", sep="/"), check.names = FALSE)
ref_marker = list()
for(r in 1:nrow(ref_marker_d)){
 gene = ref_marker_d[[1]][r]
 cell_type = ref_marker_d[[2]][r]
 if(is.null(ref_marker[[cell_type]])){
   ref_marker[[cell_type]] = c()
 }
 ref_marker[[cell_type]] = c(ref_marker[[cell_type]], gene)
}
result = RetrofitMapByMarkers(ref_marker = ref_marker, 
                             K=K,
                             decomp_w = decomp_w,
                             decomp_h = decomp_h)
}
\seealso{
papers reference
}
