% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{RetrofitMatch}
\alias{RetrofitMatch}
\title{RETROFIT matching algorithm}
\usage{
RetrofitMatch(ref_w, w, h)
}
\arguments{
\item{h}{}

\item{x}{Matrix(GeneExpressions, Spots): Spatial Transciptomics Data.}
}
\value{
A list of
\itemize{
\item w
\item h
}
}
\description{
matching algorithm description
}
\seealso{
papers reference
}
