% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testAnnotateWithCorrelationsData}
\alias{testAnnotateWithCorrelationsData}
\title{annotate with correlation test}
\format{
Includes input decomp_w, decomp_h, sc_ref and output cell types
}
\usage{
data(testAnnotateWithCorrelationsData)
}
\description{
A dataset with input&output of "annotateWithCorrelations"
}
\details{
\itemize{
\item testAnnotateWithCorrelationsData
}
}
\keyword{datasets}
