% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testAnnotateWithMarkersData}
\alias{testAnnotateWithMarkersData}
\title{annotate with markers test}
\format{
Includes input decomp_w, decomp_h, sc_ref and output cell types
}
\usage{
data(testAnnotateWithMarkersData)
}
\description{
A dataset with input&output of "annotateWithMarkers" reproducibility
}
\details{
\itemize{
\item testAnnotateWithMarkersData
}
}
\keyword{datasets}
