% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestDecomposeData}
\alias{TestDecomposeData}
\title{decompose test}
\format{
Includes input x and output w, h, and th
}
\usage{
data(TestDecomposeData)
}
\description{
A dataset with input and output of "decompose" for reproducibility tests.
}
\details{
\itemize{
\item TestDecomposeData
}
}
\keyword{datasets}
