% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testRetrofitData}
\alias{testRetrofitData}
\title{retrofit test}
\format{
Includes input x, references and results with large iterations
}
\usage{
data(testRetrofitData)
}
\description{
A dataset with input and output of "retrofit" for reproducibility tests.
}
\details{
\itemize{
\item testRetrofitData
}
}
\keyword{datasets}
