% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testRetrofitObj}
\alias{testRetrofitObj}
\title{retrofit test}
\format{
Includes input x, references and results with large iterations
}
\usage{
data(testRetrofitObj)
}
\description{
A dataset with input and output of "retrofit" for reproducibility tests.
}
\details{
\itemize{
\item testRetrofitObj
}
}
\keyword{datasets}
