% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignetteColonObj}
\alias{vignetteColonObj}
\title{colon vignette}
\format{
Includes colon scenario x, references, a large iterations results.
}
\usage{
data(vignetteColonObj)
}
\description{
A dataset supporting the colon vignette process
}
\details{
\itemize{
\item vignetteColonObj
}
}
\keyword{datasets}
