% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignetteSimulationData}
\alias{vignetteSimulationData}
\title{SimulationVignette process supporting data}
\format{
\subsection{\code{vignetteSimulationData}}{

A data frame with sample input for RETROFIT, reference for annotation purposes, decomposed results of the sample data using 4000 iterations and true cell type proportions for the sample input data for analysis.
\describe{
\item{n10m3_x}{500 x 1000 GeneExpressions x Spots}
\item{sc_ref_w}{500 x 10 GeneExpressions x Cell types}
\item{sc_ref_h}{10 x 1000 Cell types x Spots}
\item{results_4k_iterations/}{Results of 4k iterations computation}
\item{results_4k_iterations/decompose/h}{20 x 1000 Components x Spots}
\item{results_4k_iterations/decompose/w}{500 x 20 GeneExpressions x Components}
\item{results_4k_iterations/decompose/th}{20 Components}
}
}
}
\source{
Roopali Singh, Xi He, Adam Keebum Park, Ross Cameron Hardison, Xiang Zhu, Qunhua Li, RETROFIT: Reference-free deconvolution of cell-type mixtures in spatial transcriptomics, Preprint Forthcoming (2023)
}
\usage{
data(vignetteSimulationData)
}
\description{
A dataset that supports the process of following SimulationVignette.
Items include input synthetic ST data to be decomposed, and resulting data requiring expensive computations. The synthetic sample data and reference is created using scRNA-seq data published in \href{https://pubmed.ncbi.nlm.nih.gov/30923225/}{paper}.
}
\details{
\itemize{
\item vignetteSimulationData
}
}
\keyword{datasets}
