% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateWithCorrelations.R
\name{annotateWithCorrelations}
\alias{annotateWithCorrelations}
\title{RETROFIT matching algorithm}
\usage{
annotateWithCorrelations(sc_ref, K, decomp_w, decomp_h)
}
\arguments{
\item{K}{integer: The number of cell types to be selected}

\item{decomp_w}{Matrix(GeneExpressions, Components): Decomposed w matrix}

\item{decomp_h}{Matrix(Components, Spots): Decomposed h matrix}

\item{ref_w}{A Matrix or Array with two dimensions (GeneExpressions, Cell types).}
}
\value{
A list of
\itemize{
\item w
\item h
}
}
\description{
Match cell types based on correlations with reference. decomp_w   between matching algorithm description
}
\examples{
K = 8
}
\seealso{
papers reference
}
