% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{RETROFIT decomposition algorithm}
\usage{
decompose(
  x,
  L = 16,
  iterations = 4000,
  lambda = 0.01,
  seed = NULL,
  alpha_w_0 = 0.05,
  beta_w_0 = 1e-04,
  alpha_h_0 = 0.2,
  beta_h_0 = 0.2,
  alpha_th_0 = 1.25,
  beta_th_0 = 10,
  kappa = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{x}{Matrix or Array with dimension (GeneExpressions, Spots).
This is the main spatial transciptomics data.}

\item{L}{integer: The number of components to be decomposed}

\item{iterations}{integer: The number of maximum iterations to be executed}

\item{lambda}{double: Background expression profile control}

\item{seed}{double: Random variable seed in case the output should be deterministic}

\item{alpha_w_0}{double: Variational initial parameter for vector alpha_w}

\item{beta_w_0}{double:  Variational initial parameter for vector beta_w}

\item{alpha_h_0}{double: Variational initial parameter for vector alpha_h}

\item{beta_h_0}{double:  Variational initial parameter for vector beta_h}

\item{alpha_th_0}{double:Variational initial parameter for vector alpha_th}

\item{beta_th_0}{double: Variational initial parameter for vector beta_th}

\item{kappa}{double: Learning rate factor}

\item{tolerance}{double: tolerance factor for convergence of the algorithm}

\item{plot}{boolean: Plot relative errors}
}
\value{
A list of decomposed vectors that contains
\itemize{
\item w: 2d array with GeneExpressions, Components
\item h: 2d array with Components, Spots
\item th: an array with Components
}
}
\description{
Receiving the input with 2d spatial transcriptomics matrix,
the function returns factorized {W, H, θ}.
This function fulfills Structured Stochastic Variational Inference Algorithm for RETROFIT.
Since exact Bayesian inference is infeasible and considering the large number of spots and genes,
variational inference was adopted to approximately estimate the parameters in performant manner.
}
\examples{
iterations = 10
L = 16
x=read.csv(paste("../data", "sample_x.csv", sep="/"), row.names = 1, check.names = FALSE)
result = RetrofitDecompose(x,
                          iterations=iterations,
                          L=L)
}
\seealso{
papers reference
}
