% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReadmeData}
\alias{ReadmeData}
\title{Readme}
\format{
Includes input x
}
\usage{
data(ReadmeData)
}
\description{
A dataset supporting readme example
}
\details{
\itemize{
\item ReadmeData
}
}
\keyword{datasets}
