% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit.R
\name{retrofit}
\alias{retrofit}
\title{RETROFIT}
\usage{
retrofit(
  x,
  sc_ref = NULL,
  marker_ref = NULL,
  L = 16,
  K = 8,
  iterations = 4000,
  lambda = 0.01,
  seed = NULL,
  alpha_w_0 = 0.05,
  beta_w_0 = 1e-04,
  alpha_h_0 = 0.2,
  beta_h_0 = 0.2,
  alpha_th_0 = 1.25,
  beta_th_0 = 10,
  kappa = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{x}{A matrix or array with dimension (GeneExpressions, Spots). This is the main spatial transciptomics data.}

\item{sc_ref}{A matrix or array with two dimensions (GeneExpressions, Cell types).}

\item{marker_ref}{A list with (keys, values) = (cell types, an array of genes).}

\item{L}{integer (default:16)    The number of components to be decomposed}

\item{K}{integer: The number of cell types to be selected}

\item{iterations}{integer (default:4000)  The number of maximum iterations to be executed}

\item{lambda}{double  (default:0.01)  Background expression profile control}

\item{seed}{double  (default:NULL)  Random variable seed in case the output should be deterministic}

\item{alpha_w_0}{double  (default:0.05)  Variational initial parameter for vector alpha_w}

\item{beta_w_0}{double  (default:0.0001)Variational initial parameter for vector beta_w}

\item{alpha_h_0}{double  (default:0.2)   Variational initial parameter for vector alpha_h}

\item{beta_h_0}{double  (default:0.2)   Variational initial parameter for vector beta_h}

\item{alpha_th_0}{double  (default:1.25)  Variational initial parameter for vector alpha_th}

\item{beta_th_0}{double  (default:10)    Variational initial parameter for vector beta_th}

\item{kappa}{double  (default:0.5)   Learning rate factor}

\item{verbose}{boolean (default:FALSE)}
}
\value{
A list of decomposed vectors that contains
\itemize{
\item decompose: \itemize{
\item w:  Decomposed 2d array with GeneExpressions, Components
\item h:  Decomposed 2d array with Components, Spots
\item th: 1d array with Components
}
\item annotateWithCorrelations: \itemize{
\item w:  Filtered 2d array with GeneExpressions, Cell types
\item h:  Filtered2d array with Cell types, Spots
}
\item annotateWithMarkers: \itemize{
\item w:  Filtered 2d array with GeneExpressions, Cell types
\item h:  Filtered2d array with Cell types, Spots
}
}
}
\description{
The main algorithm
}
\examples{
iterations = 10
L = 16
K = 8
}
\seealso{
papers reference
}
