% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit-package.R, R/retrofit.R
\docType{package}
\name{retrofit}
\alias{retrofit}
\alias{_PACKAGE}
\alias{retrofit-package}
\title{retrofit: Reference-free deconvolution of cell mixtures in spatial transcriptomics}
\usage{
retrofit(
  x,
  sc_ref = NULL,
  marker_ref = NULL,
  L = 16,
  K = 8,
  iterations = 4000,
  init_param = NULL,
  lambda = 0.01,
  kappa = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{x}{A matrix or array with dimension (GeneExpressions, Spots). This is the main spatial transciptomics data.}

\item{sc_ref}{A matrix or array with two dimensions (GeneExpressions, Cell types).}

\item{marker_ref}{A list with (keys, values) = (cell types, an array of genes).}

\item{L}{integer (default:16)    The number of components to be decomposed}

\item{K}{integer: The number of cell types to be selected}

\item{iterations}{integer (default:4000)  The number of maximum iterations to be executed}

\item{init_param}{list                    Vatirational initial parameters}

\item{lambda}{double  (default:0.01)  Background expression profile control}

\item{kappa}{double  (default:0.5)   Learning rate factor}

\item{verbose}{boolean (default:FALSE)}
}
\value{
A list of decomposed vectors that contains
\itemize{
\item decompose: \itemize{
\item w:  Decomposed 2d array with GeneExpressions, Components
\item h:  Decomposed 2d array with Components, Spots
\item th: 1d array with Components
}
\item annotateWithCorrelations: \itemize{
\item w:  Filtered 2d array with GeneExpressions, Cell types
\item h:  Filtered2d array with Cell types, Spots
}
\item annotateWithMarkers: \itemize{
\item w:  Filtered 2d array with GeneExpressions, Cell types
\item h:  Filtered2d array with Cell types, Spots
}
}
}
\description{
RETROFIT is a Bayesian non-negative matrix factorization framework to decompose cell type mixtures in ST data without using external single-cell expression references. RETROFIT outperforms existing reference-based methods in estimating cell type proportions and reconstructing gene expressions in simulations with varying spot size and sample heterogeneity, irrespective of the quality or availability of the single-cell reference. RETROFIT recapitulates known cell-type localization patterns in a Slide-seq dataset of mouse cerebellum without using any single-cell data.

The function "retrofit" is an implementation of the main algorithm of RETROFIT.
"retrofit" is composed of two functions: decompose & annotate.
Two types of annotation is defined based on referencing method: annotate with correlations, annotate with markers.
}
\examples{
utils::data("testSimulationData")
iterations  <- 10
L           <- 16
K           <- 8
x           <- testSimulationData$extra5_x
sc_ref      <- testSimulationData$sc_ref

res         <- retrofit::retrofit(x, sc_ref=sc_ref, L=L, K=K, iterations=iterations)
W           <- res$decompose$w
W_annotated <- res$annotateWithCorrelations$w
ranked_cells<- res$annotateWithCorrelations$ranked_cells
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qunhualilab/retrofit}
  \item Report bugs at \url{https://github.com/qunhualilab/retrofit/issues}
}


papers reference
}
\author{
\strong{Maintainer}: Adam Park \email{akp6031@psu.edu}

Authors:
\itemize{
  \item Roopali Singh \email{rus82@psu.edu} (\href{https://orcid.org/0000-0001-6539-6622}{ORCID})
  \item Xiang Zhu \email{xiangzhu@psu.edu} (\href{https://orcid.org/0000-0003-1134-6413}{ORCID})
  \item Qunhua Li \email{qunhua.li@psu.edu} (\href{https://orcid.org/0000-0003-0675-7648}{ORCID})
}

}
