% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit.R
\name{retrofit}
\alias{retrofit}
\title{Split a string}
\usage{
retrofit(x, iterations = 4000)
}
\arguments{
\item{x}{Spatial Transciptomics Data. Matrix(GeneExpressions, Spots)}
}
\value{
Factorization results.
W: Matrix(GeneExpressions, Components)
H: Matrix(Components, Spots)
$\theta$: Vector(Components)
}
\description{
Split a string
}
