% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit.R
\name{retrofit}
\alias{retrofit}
\title{RETROFIT decomposition algorithm}
\usage{
retrofit(x, iterations = 4000)
}
\arguments{
\item{x}{Matrix(GeneExpressions, Spots): Spatial Transciptomics Data.}

\item{iterations}{integer: The number of iterations to be executed}
}
\value{
A list of decomposed vectors that contains
\itemize{
\item w: 2d array with GeneExpressions, Components
\item h: 2d array with Components, Spots
\item th: an array with Components
}
}
\description{
Receiving the input with 2d spatial transcriptomics matrix,
the function returns factorized {W, H, θ}.
This function fulfills Structured Stochastic Variational Inference Algorithm for RETROFIT.
Since exact Bayesian inference is infeasible and considering the large number of spots and genes,
variational inference was adopted to approximately estimate the parameters in performant manner.
}
\examples{
x=read.csv(in_path)
rownames(x)=x[,1]
x=as.matrix(x[,-1])
result = retrofit_decompose(x)
}
\seealso{
papers reference
}
