% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RetrofitMain}
\alias{RetrofitMain}
\title{RETROFIT}
\usage{
RetrofitMain(x)
}
\arguments{
\item{x}{Matrix(GeneExpressions, Spots): Spatial Transciptomics Data.}
}
\value{
A list of paths of output data that contain
\itemize{
\item w: 2d array with GeneExpressions, Components
\item h: 2d array with Components, Spots
\item th: an array with Components
\item match: an array with Components
}
}
\description{
The main algorithm
}
\examples{
temporary
}
\seealso{
papers reference
}
