% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_by_markers.R
\name{RetrofitMapByMarkers}
\alias{RetrofitMapByMarkers}
\title{RETROFIT matching algorithm}
\usage{
RetrofitMapByMarkers(ref_marker, K, decomp_w, decomp_h)
}
\arguments{
\item{x}{Matrix(GeneExpressions, Spots): Spatial Transciptomics Data.}

\item{h}{}
}
\value{
A list of
\itemize{
\item w
\item h
}
}
\description{
Match cell types based on correlations with reference. decomp_w   between matching algorithm description
}
\seealso{
papers reference
}
