% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{RetrofitMatch}
\alias{RetrofitMatch}
\title{RETROFIT matching algorithm}
\usage{
RetrofitMatch(ref_w, decomp_w, decomp_h, K)
}
\arguments{
\item{x}{Matrix(GeneExpressions, Spots): Spatial Transciptomics Data.}

\item{h}{}
}
\value{
A list of
\itemize{
\item w
\item h
}
}
\description{
matching algorithm description
}
\seealso{
papers reference
}
