% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testSimulationData}
\alias{testSimulationData}
\title{simulation data}
\format{
Includes input x, references and results with large iterations
}
\usage{
data(testSimulationData)
}
\description{
A dataset with input and output of retrofit functions for reproducibility tests.
}
\details{
\itemize{
\item testSimulationData
}
}
\keyword{datasets}
