% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VignetteColonData}
\alias{VignetteColonData}
\title{colon vignette}
\format{
Includes colon scenario x, references, a large iterations results.
}
\usage{
data(VignetteColonData)
}
\description{
A dataset supporting the colon vignette process
}
\details{
\itemize{
\item VignetteColonData
}
}
\keyword{datasets}
