% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignetteSimulationData}
\alias{vignetteSimulationData}
\title{simulation vignette}
\format{
Includes n10m3 scenario x, references, a large iterations results.
}
\usage{
data(vignetteSimulationData)
}
\description{
A dataset supporting the simulation vignette process
}
\details{
\itemize{
\item vignetteSimulationData
}
}
\keyword{datasets}
